move immediate value helpers to attrs file
